+++ src/config.rs
@@ -1,4 +1,5 @@
 use std::{
+  collections::HashMap,
   fs,
   path::{Path, PathBuf},
 };
@@ -8,7 +9,8 @@ use crossterm::style::Stylize;
 use derive_more::{Display, IsVariant};
 #[cfg(test)]
 use fake::{Dummy, Fake};
-use miette::{Diagnostic, Result};
+use miette::{Diagnostic, NamedSource, Result, SourceSpan};
+use path_absolutize::Absolutize;
 use serde::{Deserialize, Serialize};
 use somok::Somok;
 
@@ -23,6 +25,22 @@ pub enum LinkType {
   Hard,
 }
 
+#[cfg(test)]
+struct ValueFaker;
+
+#[cfg(test)]
+impl Dummy<ValueFaker> for HashMap<String, serde_json::Value> {
+  fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &ValueFaker, rng: &mut R) -> Self {
+    let mut map = HashMap::new();
+
+    for _ in 0..10.fake_with_rng(rng) {
+      map.insert((0..10).fake_with_rng(rng), serde_json::Value::String((0..10).fake_with_rng::<String, R>(rng)));
+    }
+
+    map
+  }
+}
+
 #[derive(Deserialize, Serialize, Debug)]
 #[cfg_attr(test, derive(Dummy, PartialEq))]
 pub struct Config {
@@ -32,13 +50,14 @@ pub struct Config {
   /// Which link type to use for linking dotfiles
   pub(crate) link_type: LinkType,
 
-  /// The url of the repository passed to the git clone command
-  pub(crate) repo: Option<String>,
-
   /// The command used to spawn processess.
   /// Use handlebars templates `{{ cmd }}` as placeholder for the cmd set in the dot.
   /// E.g. `"bash -c {{ quote "" cmd }}"`.
   pub(crate) shell_command: Option<String>,
+
+  /// Variables can be used for templating in dot.(yaml|toml|json) files.
+  #[cfg_attr(test, dummy(faker = "ValueFaker"))]
+  pub(crate) variables: HashMap<String, serde_json::Value>,
 }
 
 impl Default for Config {
@@ -46,13 +65,13 @@ impl Default for Config {
     Self {
       dotfiles: USER_DIRS.home_dir().join(".dotfiles"),
       link_type: LinkType::Symbolic,
-      repo: None,
       #[cfg(windows)]
       shell_command: Some("pwsh -NoProfile -C {{ quote \"\" cmd }}".to_string()),
       #[cfg(all(not(target_os = "macos"), unix))]
       shell_command: Some("bash -c {{ quote \"\" cmd }}".to_string()),
       #[cfg(target_os = "macos")]
       shell_command: Some("zsh -c {{ quote \"\" cmd }}".to_string()),
+      variables: HashMap::new(),
     }
   }
 }
@@ -63,8 +82,35 @@ fn deserialize_config(config: &str) -> Result<Config, serde_yaml::Error> {
 }
 
 #[cfg(feature = "yaml")]
-fn serialize_config(config: &Config) -> Result<String, serde_yaml::Error> {
+fn serialize(config: &impl Serialize) -> Result<String, serde_yaml::Error> {
   serde_yaml::to_string(&config)
+}
+
+#[derive(thiserror::Error, Diagnostic, Debug)]
+#[error("{name} is already set")]
+#[diagnostic(code(config::exists::value))]
+pub struct AlreadyExistsError {
+  name: String,
+  #[label("{name} is set here")]
+  span: SourceSpan,
+}
+
+impl AlreadyExistsError {
+  pub fn new(name: &str, content: &str) -> Self {
+    let pat = format!("{name}: ");
+    let span: SourceSpan = if content.starts_with(&pat) {
+      (0, pat.len()).into()
+    } else {
+      let starts = content.match_indices(&format!("\n{pat}")).collect::<Vec<_>>();
+      if starts.len() == 1 {
+        (starts[0].0 + 1, pat.len()).into()
+      } else {
+        (0, content.len()).into()
+      }
+    };
+
+    Self { name: name.to_string(), span }
+  }
 }
 
 #[derive(thiserror::Error, Diagnostic, Debug)]
@@ -77,24 +123,58 @@ pub enum Error {
   #[error("Could not write config")]
   #[diagnostic(code(config::write))]
   WritingConfig(PathBuf, #[source] std::io::Error),
+
+  #[error("Could not get absolute path")]
+  #[diagnostic(code(config::canonicalize))]
+  Canonicalize(#[source] std::io::Error),
+
+  #[error("Config file already exists")]
+  #[diagnostic(code(config::exists))]
+  AlreadyExists(#[label] Option<SourceSpan>, #[source_code] NamedSource, #[related] Vec<AlreadyExistsError>),
+
+  #[error("Could not parse dotfiles directory \"{0}\"")]
+  #[diagnostic(code(config::filename::parse), help("Did you enter a valid file?"))]
+  PathParse(PathBuf),
 }
 
 #[cfg_attr(all(nightly, coverage), no_coverage)]
-pub fn create_config_file_with_repo(repo: &str, config_file: &Path) -> Result<()> {
-  let mut config = Config::default();
+pub fn create_config_file(dotfiles: Option<&Path>, config_file: &Path) -> Result<(), Error> {
   if let Ok(existing_config_str) = fs::read_to_string(config_file) {
     if let Ok(existing_config) = deserialize_config(&existing_config_str) {
-      if existing_config.repo.as_ref().map_or(false, |r| *r != *repo) {
-        println!("Warning: {}", "Config file already exists and contains a different repo".yellow());
-        return ().okay();
+      let mut errors: Vec<AlreadyExistsError> = vec![];
+
+      if let Some(dotfiles) = dotfiles {
+        if existing_config.dotfiles != dotfiles {
+          errors.push(AlreadyExistsError::new("dotfiles", &existing_config_str));
+        }
       }
-      config = existing_config;
+
+      return Error::AlreadyExists(
+        errors.is_empty().then(|| (0, existing_config_str.len()).into()),
+        NamedSource::new(config_file.to_string_lossy(), existing_config_str),
+        errors,
+      )
+      .error();
     }
   }
 
-  config.repo = repo.to_string().some();
+  let mut map = HashMap::new();
 
-  fs::write(config_file, serialize_config(&config).map_err(Error::SerializingConfig)?).map_err(|e| Error::WritingConfig(config_file.to_path_buf(), e))?;
+  if let Some(dotfiles) = dotfiles {
+    map.insert(
+      "dotfiles",
+      dotfiles
+        .absolutize()
+        .map_err(Error::Canonicalize)?
+        .to_str()
+        .ok_or_else(|| Error::PathParse(dotfiles.to_path_buf()))?
+        .to_string(),
+    );
+  }
+
+  fs::write(config_file, serialize(&map).map_err(Error::SerializingConfig)?).map_err(|e| Error::WritingConfig(config_file.to_path_buf(), e))?;
+
+  println!("Created config file at {}", config_file.to_string_lossy().green());
 
   ().okay()
 }
@@ -111,7 +191,7 @@ mod tests {
   #[case(Faker.fake::<Config>())]
   #[case(Config::default())]
   fn ser_de(#[case] config: Config) {
-    let serialized = super::serialize_config(&config);
+    let serialized = super::serialize(&config);
     assert_that!(&serialized).is_ok();
     let serialized = serialized.unwrap();